Ported most of KubeJS to 1.14. Now supports datapack "mods" and /reload. Un-reloadable scripts that are only loaded once at startup now go in kubejs/startup/. scripts.json is required and will be auto-generated

+++ src/main/java/dev/latvian/kubejs/entity/CheckLivingEntitySpawnEventJS.java
@@ -35,17 +35,17 @@ public class CheckLivingEntitySpawnEventJS extends LivingEntityEventJS
 		return entityOf(event);
 	}
 
-	public float getX()
+	public double getX()
 	{
 		return event.getX();
 	}
 
-	public float getY()
+	public double getY()
 	{
 		return event.getY();
 	}
 
-	public float getZ()
+	public double getZ()
 	{
 		return event.getZ();
 	}
